+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - New `extract-base64` command to extract and decode base64 strings from events. (#1512) (@fukusuket)
 - New `expand-list` command to output placeholder names used for rules with the `expand` modifier. (#1513) (@fukuseket)
 - Support for `expand` field modifiers. (#1434) (@fukusuket)
+- Suppport for Temporal Proximity (`temporal`) correlation rules. (#1446) (@fukusuket)
 
 **Enhancements:**